test(coderd/x/chatd): drop chain mode test coverage
