+++ coderd/x/chatd/chatd_test.go
@@ -12500,199 +12500,6 @@ func TestAdvisorGating_ExploreSubagent(t *testing.T) {
 	}
 }
 
-// TestAdvisorChainMode_SnapshotKeepsFullHistory exercises the advisor
-// runtime together with chain mode and asserts the snapshot captured for
-// the nested advisor call retains the full pre-chain prompt. Chain mode
-// otherwise strips assistant and tool turns from the prompt the outer
-// loop sees, so a regression that captures the advisor snapshot after
-// filterPromptForChainMode, or removes the chain-mode guard around
-// advisor snapshotting, would leak the filtered view into the advisor's
-// nested call. The advisor would then only see the trailing user
-// message, losing the context the outer model had been building on.
-func TestAdvisorChainMode_SnapshotKeepsFullHistory(t *testing.T) {
-	t.Parallel()
-
-	db, ps := dbtestutil.NewDB(t)
-	ctx := testutil.Context(t, testutil.WaitLong)
-
-	const (
-		turn1User    = "help me refactor this module"
-		turn1Reply   = "happy to help, tell me more"
-		turn1RespID  = "resp_turn1_advisor_chain"
-		turn2User    = "follow up question"
-		advisorReply = "narrow the scope to one module"
-		finalReply   = "acknowledged"
-	)
-
-	var (
-		requestsMu        sync.Mutex
-		requests          []recordedOpenAIRequest
-		advisorRequestRaw []byte
-		advisorCallSeen   atomic.Bool
-	)
-
-	openAIURL := chattest.NewOpenAI(t, func(req *chattest.OpenAIRequest) chattest.OpenAIResponse {
-		if !req.Stream {
-			return chattest.OpenAINonStreamingResponse("title")
-		}
-
-		// The advisor's nested call runs with no tools (MaxSteps=1,
-		// empty tool set). Parent calls always carry the chat's tool
-		// set, which includes the advisor tool.
-		isAdvisorNested := len(req.Tools) == 0
-
-		requestsMu.Lock()
-		requests = append(requests, recordOpenAIRequest(req))
-		if isAdvisorNested {
-			advisorRequestRaw = append([]byte(nil), req.RawBody...)
-			advisorCallSeen.Store(true)
-		}
-		requestsMu.Unlock()
-
-		if isAdvisorNested {
-			return chattest.OpenAIStreamingResponse(
-				chattest.OpenAITextChunks(advisorReply)...,
-			)
-		}
-
-		// Turn 1 parent request: no previous_response_id yet, so chain
-		// mode cannot activate. Respond with a plain text reply and
-		// tag the stored response id so turn 2 can chain off it.
-		if req.PreviousResponseID == nil {
-			resp := chattest.OpenAIStreamingResponse(
-				chattest.OpenAITextChunks(turn1Reply)...,
-			)
-			resp.ResponseID = turn1RespID
-			return resp
-		}
-
-		// Turn 2 parent: chain mode is active. On the first pass call
-		// advisor; on the continuation after the tool result arrives,
-		// close out with a final text reply.
-		var hasAdvisorResult bool
-		for _, m := range req.Messages {
-			if m.Role == "tool" && strings.Contains(m.Content, advisorReply) {
-				hasAdvisorResult = true
-				break
-			}
-		}
-		if !hasAdvisorResult {
-			return chattest.OpenAIStreamingResponse(chattest.OpenAIToolCallChunk(
-				"advisor",
-				`{"question":"should I keep going?"}`,
-			))
-		}
-		return chattest.OpenAIStreamingResponse(
-			chattest.OpenAITextChunks(finalReply)...,
-		)
-	})
-
-	user, org, _ := seedChatDependenciesWithProvider(t, db, "openai", openAIURL)
-	storeEnabled := true
-	// The OpenAI Responses API is the only provider code path where
-	// chain mode activates. Store=true is the switch that routes this
-	// provider/model through the Responses API and lets
-	// IsResponsesStoreEnabled return true.
-	responsesModel := insertChatModelConfigWithCallConfig(
-		t, db, user.ID, "openai", "gpt-4o",
-		codersdk.ChatModelCallConfig{
-			ProviderOptions: &codersdk.ChatModelProviderOptions{
-				OpenAI: &codersdk.ChatModelOpenAIProviderOptions{
-					Store: &storeEnabled,
-				},
-			},
-		},
-	)
-	seedAdvisorConfig(ctx, t, db, codersdk.AdvisorConfig{
-		Enabled:         true,
-		MaxUsesPerRun:   3,
-		MaxOutputTokens: 16384,
-	})
-	server := newOpenAIResponsesTestServer(t, db, ps, func(cfg *chatd.Config) {
-		cfg.AIBridgeTransportFactory = chatAIGatewayTransportFactoryPointer(
-			chattest.NewMockAIBridgeTransport(t, openAIURL),
-		)
-	})
-
-	chat, err := server.CreateChat(ctx, chatd.CreateOptions{
-		OrganizationID: org.ID,
-		OwnerID:        user.ID,
-		APIKeyID:       testAPIKeyID(t, db, user.ID),
-		Title:          "advisor-chain-mode",
-		ModelConfigID:  responsesModel.ID,
-		InitialUserContent: []codersdk.ChatMessagePart{
-			codersdk.ChatMessageText(turn1User),
-		},
-	})
-	require.NoError(t, err)
-
-	// Turn 1 must settle before turn 2 starts so the assistant row
-	// with ProviderResponseID is visible to resolveChainMode.
-	waitForChatProcessed(ctx, t, db, chat.ID, server)
-	turn1Chat, err := db.GetChatByID(ctx, chat.ID)
-	require.NoError(t, err)
-	require.Equal(t, database.ChatStatusWaiting, turn1Chat.Status,
-		"turn 1 must complete before turn 2 can be sent; last_error=%q", chatLastErrorMessage(turn1Chat.LastError))
-
-	_, err = server.SendMessage(ctx, chatd.SendMessageOptions{
-		ChatID:    chat.ID,
-		CreatedBy: user.ID,
-		APIKeyID:  testAPIKeyID(t, db, user.ID),
-		Content: []codersdk.ChatMessagePart{
-			codersdk.ChatMessageText(turn2User),
-		},
-	})
-	require.NoError(t, err)
-
-	require.Eventually(t, func() bool {
-		if !advisorCallSeen.Load() {
-			return false
-		}
-		got, getErr := db.GetChatByID(ctx, chat.ID)
-		if getErr != nil {
-			return false
-		}
-		return got.Status == database.ChatStatusWaiting ||
-			got.Status == database.ChatStatusError
-	}, testutil.WaitLong, testutil.IntervalFast)
-
-	requestsMu.Lock()
-	gotAdvisorBody := append([]byte(nil), advisorRequestRaw...)
-	gotRequests := append([]recordedOpenAIRequest(nil), requests...)
-	requestsMu.Unlock()
-
-	// Chain mode must have actually fired on turn 2, otherwise this
-	// test degenerates to TestAdvisorHappyPath_RootChat.
-	var chainModeActivated bool
-	for _, r := range gotRequests {
-		if r.PreviousResponseID != nil && *r.PreviousResponseID == turn1RespID {
-			chainModeActivated = true
-			break
-		}
-	}
-	require.True(t, chainModeActivated,
-		"turn 2 parent request must carry previous_response_id; without it this test does not exercise chain mode")
-
-	require.True(t, advisorCallSeen.Load(),
-		"the nested advisor call must execute under chain mode")
-	require.NotEmpty(t, gotAdvisorBody,
-		"advisor call must receive a non-empty request body")
-
-	// The core assertion: the advisor snapshot must retain turn 1
-	// context. Chain mode filtering strips assistant and tool turns
-	// from the prompt the outer loop sees, so if that filtered view
-	// leaked into the snapshot the advisor would only see turn 2's
-	// trailing user message. The advisor's nested call goes through
-	// the OpenAI Responses API, which encodes its prompt in the
-	// "input" field rather than "messages", so we inspect the raw
-	// request body for both turn-1 substrings.
-	require.Contains(t, string(gotAdvisorBody), turn1User,
-		"advisor snapshot must retain the turn 1 user message even when chain mode is active")
-	require.Contains(t, string(gotAdvisorBody), turn1Reply,
-		"advisor snapshot must retain the turn 1 assistant message even when chain mode is active")
-}
-
-// TestProviderSwitchSanitizesAndRestoresPEToolHistory verifies the A→B→A
 // provider-switch contract:
 //
 //  1. A turn using model MA (backed by provider A) produces a
