+++ coderd/x/chatd/chaterror/classify_test.go
@@ -1465,123 +1465,6 @@ func TestClassify_TruncatesProviderDetail(t *testing.T) {
 	require.True(t, strings.HasSuffix(classified.Detail, "…"))
 }
 
-func TestClassify_ChainBroken(t *testing.T) {
-	t.Parallel()
-
-	tests := []struct {
-		name            string
-		err             error
-		wantChainBroken bool
-		wantRetryable   bool
-		wantProvider    string
-		wantStatusCode  int
-	}{
-		{
-			name: "OpenAIPreviousResponseNotFoundBareString",
-			err: xerrors.New(
-				"Previous response with id 'resp_abc' not found.",
-			),
-			wantChainBroken: true,
-			wantRetryable:   true,
-			wantProvider:    "openai",
-			wantStatusCode:  0,
-		},
-		{
-			name: "OpenAIPreviousResponseNotFoundProviderError",
-			err: testProviderError(
-				"Previous response with id 'resp_096c70c5bb8d52bc0069fa11e0630c81a3ba210cddfa75bae9' not found.",
-				404,
-				nil,
-			),
-			wantChainBroken: true,
-			wantRetryable:   true,
-			wantProvider:    "openai",
-			wantStatusCode:  404,
-		},
-		{
-			name: "OpenAIPreviousResponseCaseInsensitive",
-			err: testProviderError(
-				"PREVIOUS RESPONSE WITH ID 'resp_abc' NOT FOUND.",
-				404,
-				nil,
-			),
-			wantChainBroken: true,
-			wantRetryable:   true,
-			wantProvider:    "openai",
-			wantStatusCode:  404,
-		},
-		{
-			name: "PreviousResponseWithoutNotFoundIsNotChainBroken",
-			err: testProviderError(
-				"Previous response with id 'resp_abc' is invalid.",
-				400,
-				nil,
-			),
-			wantChainBroken: false,
-		},
-		{
-			name: "UnrelatedNotFoundIsNotChainBroken",
-			err: testProviderError(
-				"resource not found",
-				404,
-				nil,
-			),
-			wantChainBroken: false,
-		},
-		{
-			name: "UnrelatedInvalidRequestIsNotChainBroken",
-			err: testProviderError(
-				"",
-				400,
-				nil,
-				testProviderResponseDump(`{"error":{"type":"invalid_request_error","message":"Image exceeds 5 MB maximum."}}`),
-			),
-			wantChainBroken: false,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			t.Parallel()
-
-			classified := chaterror.Classify(tt.err)
-			require.Equal(t, tt.wantChainBroken, classified.ChainBroken,
-				"chain broken flag mismatch")
-			if !tt.wantChainBroken {
-				return
-			}
-			require.Equal(t, tt.wantRetryable, classified.Retryable,
-				"chain-broken errors must be retryable so the loop"+
-					" can self-heal")
-			require.Equal(t, tt.wantProvider, classified.Provider)
-			require.Equal(t, tt.wantStatusCode, classified.StatusCode)
-			require.Equal(t, codersdk.ChatErrorKindGeneric, classified.Kind,
-				"chain-broken keeps the user-visible kind unchanged"+
-					" so we don't add a new codersdk surface")
-		})
-	}
-}
-
-func TestClassify_ChainBrokenSurvivesWithClassification(t *testing.T) {
-	t.Parallel()
-
-	original := chaterror.Classify(testProviderError(
-		"Previous response with id 'resp_abc' not found.",
-		404,
-		nil,
-	))
-	require.True(t, original.ChainBroken)
-
-	wrapped := chaterror.WithClassification(
-		xerrors.New("transport blew up"),
-		original,
-	)
-	round := chaterror.Classify(wrapped)
-	require.True(t, round.ChainBroken,
-		"WithClassification round-trips ChainBroken so the retry path"+
-			" can detect it after re-classification")
-}
-
 func TestClassify_MissingKeyPreClassified(t *testing.T) {
 	t.Parallel()
