+++ coderd/x/chatd/integration_responses_test.go
@@ -2,7 +2,6 @@ package chatd_test
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"strings"
 	"sync"
@@ -13,11 +12,9 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"github.com/coder/coder/v2/coderd/database"
-	"github.com/coder/coder/v2/coderd/database/dbgen"
 	"github.com/coder/coder/v2/coderd/database/dbtestutil"
 	dbpubsub "github.com/coder/coder/v2/coderd/database/pubsub"
 	"github.com/coder/coder/v2/coderd/x/chatd"
-	"github.com/coder/coder/v2/coderd/x/chatd/chatprompt"
 	"github.com/coder/coder/v2/coderd/x/chatd/chattest"
 	"github.com/coder/coder/v2/codersdk"
 	"github.com/coder/coder/v2/testutil"
@@ -198,165 +195,6 @@ func TestOpenAIResponsesFullReplayPairsReasoningAndWebSearch(t *testing.T) {
 	requirePromptItemReferenceOrder(t, followup.Prompt, reasoningID, webSearchID)
 }
 
-func TestOpenAIResponsesChainModeSkipsWhenLocalCallPending(t *testing.T) {
-	t.Parallel()
-
-	db, ps := dbtestutil.NewDB(t)
-	ctx := testutil.Context(t, testutil.WaitLong)
-
-	var recorder responsesRequestRecorder
-	openAIURL := chattest.NewOpenAI(t, func(req *chattest.OpenAIRequest) chattest.OpenAIResponse {
-		if !req.Stream {
-			return chattest.OpenAINonStreamingResponse("title")
-		}
-		recorder.record(req)
-		resp := chattest.OpenAIStreamingResponse(
-			chattest.OpenAITextChunks("resolved after local call")...,
-		)
-		resp.ResponseID = "resp_local_pending_next"
-		return resp
-	})
-
-	user, org, _ := seedChatDependenciesWithProvider(t, db, "openai", openAIURL)
-	model := insertOpenAIResponsesModelConfig(t, db, user.ID, true, false)
-	chat := insertOpenAIResponsesChat(t, db, org.ID, user.ID, model.ID, "local-pending")
-
-	callID := fmt.Sprintf("call_local_%d", time.Now().UnixNano())
-	localCall := codersdk.ChatMessageToolCall(
-		callID,
-		"read_file",
-		json.RawMessage(`{"path":"README.md"}`),
-	)
-	insertOpenAIResponsesMessages(ctx, t, db, chat.ID, user.ID, model.ID,
-		persistedResponsesMessage{
-			role: database.ChatMessageRoleUser,
-			parts: []codersdk.ChatMessagePart{
-				codersdk.ChatMessageText("please inspect the README"),
-			},
-		},
-		persistedResponsesMessage{
-			role:               database.ChatMessageRoleAssistant,
-			parts:              []codersdk.ChatMessagePart{localCall},
-			providerResponseID: "resp_local_pending_prior",
-		},
-	)
-
-	factory := chattest.NewMockAIBridgeTransport(t, openAIURL)
-	server := newOpenAIResponsesTestServer(t, db, ps, func(cfg *chatd.Config) {
-		cfg.AIBridgeTransportFactory = chatAIGatewayTransportFactoryPointer(factory)
-	})
-	_, err := server.SendMessage(ctx, chatd.SendMessageOptions{
-		ChatID:        chat.ID,
-		CreatedBy:     user.ID,
-		APIKeyID:      testAPIKeyID(t, db, user.ID),
-		ModelConfigID: model.ID,
-		Content: []codersdk.ChatMessagePart{
-			codersdk.ChatMessageText("continue after that tool call"),
-		},
-	})
-	require.NoError(t, err)
-	waitForChatProcessed(ctx, t, db, chat.ID, server)
-	requireResponsesChatWaiting(ctx, t, db, chat.ID)
-
-	requests := recorder.all()
-	require.Len(t, requests, 1)
-	request := requests[0]
-	require.NotNil(t, request.Store)
-	require.True(t, *request.Store)
-	require.Nil(t, request.PreviousResponseID)
-	require.NotEmpty(t, request.Prompt)
-	requirePromptItemWithTypeAndCallID(t, request.Prompt, "function_call", callID)
-	requirePromptItemWithTypeAndCallID(t, request.Prompt, "function_call_output", callID)
-}
-
-func TestOpenAIResponsesChainModeStillFiresForProviderExecutedOnly(t *testing.T) {
-	t.Parallel()
-
-	db, ps := dbtestutil.NewDB(t)
-	ctx := testutil.Context(t, testutil.WaitLong)
-
-	var recorder responsesRequestRecorder
-	openAIURL := chattest.NewOpenAI(t, func(req *chattest.OpenAIRequest) chattest.OpenAIResponse {
-		if !req.Stream {
-			return chattest.OpenAINonStreamingResponse("title")
-		}
-		recorder.record(req)
-		resp := chattest.OpenAIStreamingResponse(
-			chattest.OpenAITextChunks("chained answer")...,
-		)
-		resp.ResponseID = "resp_provider_only_next"
-		return resp
-	})
-
-	user, org, _ := seedChatDependenciesWithProvider(t, db, "openai", openAIURL)
-	model := insertOpenAIResponsesModelConfig(t, db, user.ID, true, true)
-	chat := insertOpenAIResponsesChat(t, db, org.ID, user.ID, model.ID, "provider-only")
-
-	const (
-		previousResponseID = "resp_provider_only_prior"
-		webSearchID        = "ws_provider_only_search"
-	)
-	webSearchCall := codersdk.ChatMessageToolCall(
-		webSearchID,
-		"web_search",
-		json.RawMessage(`{"query":"coder docs"}`),
-	)
-	webSearchCall.ProviderExecuted = true
-	webSearchResult := codersdk.ChatMessageToolResult(
-		webSearchID,
-		"web_search",
-		json.RawMessage(`{"status":"completed"}`),
-		false,
-		false,
-	)
-	webSearchResult.ProviderExecuted = true
-	insertOpenAIResponsesMessages(ctx, t, db, chat.ID, user.ID, model.ID,
-		persistedResponsesMessage{
-			role: database.ChatMessageRoleUser,
-			parts: []codersdk.ChatMessagePart{
-				codersdk.ChatMessageText("look up the docs"),
-			},
-		},
-		persistedResponsesMessage{
-			role: database.ChatMessageRoleAssistant,
-			parts: []codersdk.ChatMessagePart{
-				webSearchCall,
-				webSearchResult,
-			},
-			providerResponseID: previousResponseID,
-		},
-	)
-
-	factory := chattest.NewMockAIBridgeTransport(t, openAIURL)
-	server := newOpenAIResponsesTestServer(t, db, ps, func(cfg *chatd.Config) {
-		cfg.AIBridgeTransportFactory = chatAIGatewayTransportFactoryPointer(factory)
-	})
-	_, err := server.SendMessage(ctx, chatd.SendMessageOptions{
-		ChatID:        chat.ID,
-		CreatedBy:     user.ID,
-		APIKeyID:      testAPIKeyID(t, db, user.ID),
-		ModelConfigID: model.ID,
-		Content: []codersdk.ChatMessagePart{
-			codersdk.ChatMessageText("what did it find"),
-		},
-	})
-	require.NoError(t, err)
-	waitForChatProcessed(ctx, t, db, chat.ID, server)
-	requireResponsesChatWaiting(ctx, t, db, chat.ID)
-
-	requests := recorder.all()
-	require.Len(t, requests, 1)
-	request := requests[0]
-	require.NotNil(t, request.Store)
-	require.True(t, *request.Store)
-	require.NotNil(t, request.PreviousResponseID)
-	require.Equal(t, previousResponseID, *request.PreviousResponseID)
-	require.NotEmpty(t, request.Prompt)
-	requireNoResponsesProviderItemReplay(t, request.Prompt, webSearchID)
-	require.NotContains(t, promptItemTypes(request.Prompt), "web_search_call")
-	require.NotContains(t, promptItemRoles(request.Prompt), "assistant")
-}
-
 type recordedResponsesRequest struct {
 	Prompt             []interface{}
 	Store              *bool
@@ -394,12 +232,6 @@ func (r *responsesRequestRecorder) all() []recordedResponsesRequest {
 	r.mu.Lock()
 	defer r.mu.Unlock()
 	return append([]recordedResponsesRequest(nil), r.requests...)
-}
-
-type persistedResponsesMessage struct {
-	role               database.ChatMessageRole
-	parts              []codersdk.ChatMessagePart
-	providerResponseID string
 }
 
 func newOpenAIResponsesTestServer(
@@ -445,64 +277,6 @@ func insertOpenAIResponsesModelConfig(
 	)
 }
 
-func insertOpenAIResponsesChat(
-	t *testing.T,
-	db database.Store,
-	organizationID uuid.UUID,
-	ownerID uuid.UUID,
-	modelConfigID uuid.UUID,
-	titlePrefix string,
-) database.Chat {
-	t.Helper()
-	return dbgen.Chat(t, db, database.Chat{
-		OrganizationID:    organizationID,
-		OwnerID:           ownerID,
-		LastModelConfigID: modelConfigID,
-		Title:             uniqueResponsesTitle(t, titlePrefix),
-		Status:            database.ChatStatusWaiting,
-		MCPServerIDs:      []uuid.UUID{},
-		ClientType:        database.ChatClientTypeApi,
-	})
-}
-
-func insertOpenAIResponsesMessages(
-	ctx context.Context,
-	t *testing.T,
-	db database.Store,
-	chatID uuid.UUID,
-	createdBy uuid.UUID,
-	modelConfigID uuid.UUID,
-	messages ...persistedResponsesMessage,
-) {
-	t.Helper()
-	params := database.InsertChatMessagesParams{ChatID: chatID}
-	for _, message := range messages {
-		content, err := chatprompt.MarshalParts(message.parts)
-		require.NoError(t, err)
-		params.CreatedBy = append(params.CreatedBy, createdBy)
-		params.ModelConfigID = append(params.ModelConfigID, modelConfigID)
-		params.Role = append(params.Role, message.role)
-		params.Content = append(params.Content, string(content.RawMessage))
-		params.ContentVersion = append(params.ContentVersion, chatprompt.CurrentContentVersion)
-		params.Visibility = append(params.Visibility, database.ChatMessageVisibilityBoth)
-		params.InputTokens = append(params.InputTokens, 0)
-		params.OutputTokens = append(params.OutputTokens, 0)
-		params.TotalTokens = append(params.TotalTokens, 0)
-		params.ReasoningTokens = append(params.ReasoningTokens, 0)
-		params.CacheCreationTokens = append(params.CacheCreationTokens, 0)
-		params.CacheReadTokens = append(params.CacheReadTokens, 0)
-		params.ContextLimit = append(params.ContextLimit, 0)
-		params.Compressed = append(params.Compressed, false)
-		params.TotalCostMicros = append(params.TotalCostMicros, 0)
-		params.RuntimeMs = append(params.RuntimeMs, 0)
-		params.ProviderResponseID = append(params.ProviderResponseID, message.providerResponseID)
-	}
-	// Keep this raw because dbgen.ChatMessage inserts one message at a time,
-	// while this helper needs to preserve variadic batch insert behavior.
-	_, err := db.InsertChatMessages(ctx, params)
-	require.NoError(t, err)
-}
-
 func requireResponsesChatWaiting(
 	ctx context.Context,
 	t *testing.T,
@@ -537,47 +311,6 @@ func promptItemTypes(prompt []interface{}) []string {
 	return types
 }
 
-func promptItemRoles(prompt []interface{}) []string {
-	roles := make([]string, 0, len(prompt))
-	for _, item := range prompt {
-		itemMap, ok := item.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		if role := chattest.StringResponseField(itemMap, "role"); role != "" {
-			roles = append(roles, role)
-		}
-	}
-	return roles
-}
-
-func requirePromptItemWithTypeAndCallID(
-	t *testing.T,
-	prompt []interface{},
-	itemType string,
-	callID string,
-) map[string]interface{} {
-	t.Helper()
-	for _, item := range prompt {
-		itemMap, ok := item.(map[string]interface{})
-		if !ok {
-			continue
-		}
-		if chattest.StringResponseField(itemMap, "type") == itemType &&
-			chattest.StringResponseField(itemMap, "call_id") == callID {
-			return itemMap
-		}
-	}
-	promptJSON, err := json.Marshal(prompt)
-	require.NoError(t, err)
-	require.FailNowf(t, "prompt item missing",
-		"missing type=%q call_id=%q in prompt %s", itemType, callID, promptJSON)
-	return nil
-}
-
-// requireNoResponsesProviderItemReplay rejects the explicit stale IDs and all
-// provider-managed Responses item IDs. Chain mode should rely on
-// previous_response_id, not replay rs_ or ws_ identifiers in prompt input.
 func requireNoResponsesProviderItemReplay(
 	t *testing.T,
 	prompt []interface{},
